adding support for scale when viewing the presentation
with this commit we attempt to provide a first solution to allow scaling of
dahu presentation at runtime: namely without requiring to regenerate the
presentation.

with this solution the scale factor is used to resize all the object at
runtime and also adjust the animations.

the main drawback of this approach is that tooltip can span outside of
your scaled presentation when the scaling factor is to small... this
open a new discussion: should we display the tooltip in the slides
are should we display them outside: like as a footnote
(check sublimetext screencast for an example). this solution would
be better to make dahu more responsive, e.g. depending on the
screen size the tooltips will be displayed at different places.