adding support for scale when viewing the presentation
with this commit we attempt to provide a second solution to allow
scaling dahu presentation at runtime: namely without requiring to
regenerate the presentation.

with this solution the scale factor is used to apply css3 transform
to the presentation container at runtime.

the main drawback of this approach is that css3 transform are not
well supported by all browser. Also when scaling with a small
scale factor the tooltips might become unreadable!
This open the same discussion as the first solution: should we
display the tooltip in the slides are should we display them outside:
like as a footnote (check sublimetext screencast for an example) that
will be in another container. This solution would be better to make
dahu more responsive, e.g. depending on the screen size the tooltips
will be displayed at different places.